Relay and RelaySender as "relay" algorithm.
This commit moves to using the JSON command line interface for configuration
rather than TCLAP. The selection of Node creation is performed in main.

This new algorithm "relay" supports many relays to overcome the single host
capacity limitation induced by having only a single relay. Credit-based flow
control is used between the RelaySender and Relay to keep the RelaySender from
overloading the Relays. The RelaySender keeps only a single credit counter for
all the Relays. The Relay to be used is choosen with a uniform random
distribution.